finally push this gigantic refactor bodge job
took about 2 years lol
