+++ .prettierignore
@@ -1,0 +1,3 @@
+dist
+*.md
+**/package.json
