+++ LICENSE
@@ -1,5 +1,6 @@
 Copyright 2021 Filipe Calaça Barbosa
 Copyright 2022 dither
+Copyright 2023-2024 Lily Tsuru/modeco80 <lily.modeco80@protonmail.ch>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
