+++ package.json
@@ -1,18 +1,38 @@
 {
-  "name": "@lucidvm/rfb",
+  "name": "@computernewb/nodejs-rfb",
   "version": "0.3.0",
   "description": "Pure nodejs VNC (RFC 6143 / RFB Protocol) client, based on vnc-rfb-client",
-  "main": "index.js",
-  "author": "Filipe Calaça Barbosa, ayunami2000, dither",
+  "author": "Filipe Calaça Barbosa, ayunami2000, dither, modeoo80",
   "license": "MIT",
+  "exports": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "type": "module",
+  "scripts": {
+    "build": "parcel build src/index.ts --target node --target types"
+  },
+  "targets": {
+    "types": {},
+    "node": {
+      "context": "node",
+      "isLibrary": true,
+      "outputFormat": "esmodule"
+    }
+  },
   "dependencies": {},
-  "devDependencies": {},
+  "devDependencies": {
+    "@parcel/packager-ts": "2.12.0",
+    "@parcel/transformer-typescript-types": "2.12.0",
+    "@types/node": "^20.12.7",
+    "parcel": "^2.12.0",
+    "prettier": "^3.2.5",
+    "typescript": ">=3.0.0"
+  },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/lucidvm/rfb.git"
+    "url": "git+https://github.com/computernewb/nodejs-rfb.git"
   },
   "bugs": {
-    "url": "https://github.com/lucidvm/rfb/issues"
+    "url": "https://github.com/computernewb/nodejs-rfb/issues"
   },
-  "homepage": "https://github.com/lucidvm/rfb#readme"
+  "homepage": "https://github.com/computernewb/nodejs-rfb#readme"
 }
