+++ src/index.ts
@@ -1,0 +1,3 @@
+/// <reference types="node" />
+
+export * from './client.js';
