+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+      "target": "ES2022",
+      "module": "ES2022",
+      "moduleResolution": "Node",
+	  "allowSyntheticDefaultImports": true,
+      "strict": true,
+    }
+}
